Encapsulate action-group decision in RecordingController
Move the playwright-cli detection and the decision to open a recording
action group out of ShellCommandExecutor and into
RecordingController::maybe_begin_action_group, so the executor no longer
leaks a recording implementation detail.
